pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cryptolabs-proxy"
version = "1.0.0"
description = "Unified reverse proxy and fleet management landing page for CryptoLabs products"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "CryptoLabs", email = "support@cryptolabs.co.za"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "questionary>=2.0",
    "jinja2>=3.0",
    "pyyaml>=6.0",
    "requests>=2.28",
]

[project.scripts]
cryptolabs-proxy = "cryptolabs_proxy.cli:main"

[project.urls]
Homepage = "https://github.com/cryptolabsza/cryptolabs-proxy"
Documentation = "https://cryptolabs.co.za/docs"
Repository = "https://github.com/cryptolabsza/cryptolabs-proxy"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cryptolabs_proxy = ["templates/*", "templates/**/*"]
